Amélioration de la page Improvements: logique du score d'opportunité, affichage des barres et corrections diverses
- Simplification du calcul du score d'opportunité basé sur le nombre de posts négatifs pondérés par pertinence
- Affichage du nombre de posts négatifs directement sur les barres avec texte explicite
- Correction de la synchronisation des filtres de période (Distribution by Product se met à jour)
- Correction de la cohérence des compteurs de posts entre la liste et le drawer (filtrage par sentiment)
- Suppression complète des références v2 obsolètes
- Mise à jour du tooltip explicatif pour la nouvelle logique du score

+++ backend/app/routers/dashboard/models.py
@@ -21,7 +21,7 @@ class PainPointsResponse(BaseModel):
 class ProductOpportunity(BaseModel):
     """Model for product opportunity score."""
     product: str = Field(..., description="Product name")
-    opportunity_score: int = Field(..., description="Opportunity score (0-100)")
+    opportunity_score: float = Field(..., description="Opportunity score (0-100)")
     negative_posts: int = Field(..., description="Number of negative posts")
     total_posts: int = Field(..., description="Total posts for this product")
     color: str = Field(..., description="Color for visualization")
